feat(chat): implement chat state persistence actions
- Add functions to manage last conversation ID
- Add functions to manage last used model ID
- Add functions to manage thinking mode state
- Refactor imports in useChat and model-loader files
